+++ tests/runners/default-features.sh
@@ -1,3 +1,13 @@
 #!/bin/bash
 
-REDIS_USERNAME=foo REDIS_PASSWORD=bar cargo test --release --lib --tests -- --test-threads=1 -- "$@"
+declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD" "REDIS_SENTINEL_PASSWORD")
+
+for env in "${arr[@]}"
+do
+  if [ -z "$env" ]; then
+    echo "$env must be set. Run `source tests/environ` if needed."
+    exit 1
+  fi
+done
+
+cargo test --release --lib --tests -- --test-threads=1 "$@"