+++ tests/runners/sentinel-features.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
 
-export RUST_BACKTRACE=full REDIS_USERNAME=foo REDIS_PASSWORD=bar REDIS_SENTINEL_PASSWORD=baz
-cargo test --release --features "sentinel-tests sentinel-auth" --lib --tests -- --test-threads=1
+declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD" "REDIS_SENTINEL_PASSWORD")
+
+for env in "${arr[@]}"
+do
+  if [ -z "$env" ]; then
+    echo "$env must be set. Run `source tests/environ` if needed."
+    exit 1
+  fi
+done
+
+cargo test --release --features "network-logs sentinel-tests sentinel-auth" --lib --tests -- --test-threads=1 "$@"