+++ tests/scripts/install_redis_centralized.sh
@@ -7,10 +7,15 @@ function check_root_dir {
   fi
 }
 
-if [ -z "$REDIS_VERSION" ]; then
-    echo "REDIS_VERSION must be set!"
+declare -a arr=("REDIS_VERSION" "REDIS_USERNAME" "REDIS_PASSWORD" "REDIS_SENTINEL_PASSWORD")
+
+for env in "${arr[@]}"
+do
+  if [ -z "$env" ]; then
+    echo "$env must be set. Run `source tests/environ` if needed."
     exit 1
-fi
+  fi
+done
 
 ROOT=$PWD
 [[ -z "${JOBS}" ]] && PARALLEL_JOBS='2' || PARALLEL_JOBS="${JOBS}"